+++ src/foundation/app.tsx
@@ -1,13 +1,11 @@
 import * as React from 'react';
 
-import { Editor } from './editor/editor';
-import { Sidebar } from './sidebar/sidebar';
+import { Editor } from '../editor/editor';
 
 export class App extends React.Component {
   render() {
     return (
       <React.Fragment>
-        <Sidebar />
         <Editor />
       </React.Fragment>
     );
