+++ src/index.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 
-import { App } from './app';
+import { App } from './foundation/app';
 
 ReactDOM.render(<App />, document.getElementById('app'));
